Kucoin-volatility-trading-bot-main/Kucoin_volatility_trading_bot_main/requirements.txt
kucoin-futures-python
PTable==0.9.2
streamlit==1.0.0
streamlit-aggrid
ccxt
protobuf~=3.19.0
altair==4
binance
matplotlib
tradingview-ta==3.3.0
colorama==0.4.6
PyYAML
requests
prettytable==3.6.0
scikit-learn
SQLAlchemy==1.4.46
pandas_ta
loguru==0.6.0
pandas==2.0.0